[Update]migratefile

+++ db/migrate/20220707063329_create_reports.rb
@@ -2,8 +2,8 @@ class CreateReports < ActiveRecord::Migration[6.1]
   def change
     create_table :reports do |t|
       
-      t.bigint :user_id
-      t.bigint :post_id
+      t.integer :user_id
+      t.integer :post_id
       t.integer :reporter_id, null: false
       t.integer :reported_id, null: false
       t.text :reason, null: false
